Add first test for slice 01 - Add Item
This adds a single test for the slice. This is an introductory commit to
get started with ReqNRoll and Gherkin style tests. I will be using
Gherkin style tests primarily to learn them and to experiment seeing
what it looks like to do Gherkin-style tests alongside an event model
that uses Given/When/Thens