Add ReqnRoll setup for ASP.NET Integration Tests
This will focuses tests on the external behavior rather than on
implementation, while keeping it closely tied to the integration test
framework to catch as many integration issues as possible. There may
still be a need to do unit tests in addition to this, but that will be
odne using other tooling.